requirements.txt
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
logging==0.4.9.6
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
azure-functions==1.23.0
